+++ app/src/main/java/com/tonkar/volleyballreferee/ui/data/game/StoredGamesListActivity.java
@@ -131,18 +131,17 @@ public class StoredGamesListActivity extends NavigationActivity implements DataS
 
     @Override
     public boolean onOptionsItemSelected(MenuItem item) {
-        switch (item.getItemId()) {
-            case R.id.action_search_games:
-                return true;
-            case R.id.action_sync:
-                updateStoredGamesList();
-                return true;
-            case R.id.action_delete_games:
-                deleteSelectedGames();
-                return true;
-            default:
-                return super.onOptionsItemSelected(item);
+        int itemId = item.getItemId();
+        if (itemId == R.id.action_search_games) {
+            return true;
+        } else if (itemId == R.id.action_sync) {
+            updateStoredGamesList();
+            return true;
+        } else if (itemId == R.id.action_delete_games) {
+            deleteSelectedGames();
+            return true;
         }
+        return super.onOptionsItemSelected(item);
     }
 
     @Override
