+++ app/src/main/java/com/tonkar/volleyballreferee/ui/data/rules/StoredRulesViewActivity.java
@@ -106,16 +106,15 @@ public class StoredRulesViewActivity extends AppCompatActivity {
 
     @Override
     public boolean onOptionsItemSelected(MenuItem item) {
-        switch (item.getItemId()) {
-            case android.R.id.home:
-                backToList();
-                return true;
-            case R.id.action_delete_rules:
-                deleteRules();
-                return true;
-            default:
-                return super.onOptionsItemSelected(item);
+        int itemId = item.getItemId();
+        if (itemId == android.R.id.home) {
+            backToList();
+            return true;
+        } else if (itemId == R.id.action_delete_rules) {
+            deleteRules();
+            return true;
         }
+        return super.onOptionsItemSelected(item);
     }
 
     @Override
