+++ app/src/main/java/com/tonkar/volleyballreferee/ui/data/team/StoredTeamViewActivity.java
@@ -99,16 +99,15 @@ public class StoredTeamViewActivity extends AppCompatActivity {
 
     @Override
     public boolean onOptionsItemSelected(MenuItem item) {
-        switch (item.getItemId()) {
-            case android.R.id.home:
-                backToList();
-                return true;
-            case R.id.action_delete_team:
-                deleteTeam();
-                return true;
-            default:
-                return super.onOptionsItemSelected(item);
+        int itemId = item.getItemId();
+        if (itemId == android.R.id.home) {
+            backToList();
+            return true;
+        } else if (itemId == R.id.action_delete_team) {
+            deleteTeam();
+            return true;
         }
+        return super.onOptionsItemSelected(item);
     }
 
     @Override
